requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
